// src/app/shared/data-access/exercises/interfaces/exercise-actions.interface.ts
export enum ExerciseActions {
  FETCH_EXERCISES = '[Exercises] Fetch exercises',
  FETCH_EXERCISES_SUCCESS = '[Exercises] Fetch exercises success',
  FETCH_EXERCISES_FAILURE = '[Exercises] Fetch exercises failure',
  FETCH_EXERCISE = '[Exercises] Fetch exercise',
  FETCH_EXERCISE_SUCCESS = '[Exercises] Fetch exercise success',
  FETCH_EXERCISE_FAILURE = '[Exercises] Fetch exercise failure',
  EXERCISE_PAGE_CHANGE = '[Exercises] Change page',
  EXERCISE_SEARCH_TERM_CHANGE = '[Exercises] Change search term',
  EXERCISE_ADVANCED_SEARCH_VALUES_CHANGE = '[Exercises] Change advanced search values',
}